refactor: move `__init__.py` to `_cli.py` for `ape_run`
